1/25 SOL 별점 완료

+++ src/main/java/movie/MovieGradeController.java
@@ -19,37 +19,44 @@ public class MovieGradeController extends HttpServlet {
 	
 	@Override
 	protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+
+		String mode = req.getParameter("mode");
 		
-		if(req.getParameter("id")==null) {
+		if(mode.equals("insert") && req.getParameter("id")==null) {
 			JSFunction.alertLocation(resp, "로그인이 필요한 서비스입니다.", "./Login/LoginForm.jsp");
 		}
 		
-		else {
-			String mode = req.getParameter("mode");
-			String idx = req.getParameter("idx");
-			int grade = Integer.parseInt(req.getParameter("reviewStar"));
-			String mcomment = req.getParameter("mcomment");
-			String id = req.getParameter("id");
+		else  {
 			
+			String idx = req.getParameter("idx");
+			String id = req.getParameter("id");
+
 			LogDTO ldto = new LogDTO();
+			LogDAO ldao = new LogDAO();
 			
 			ldto.setId(id);
 			ldto.setIdx(idx);
-			ldto.setGrade(grade);
-			ldto.setMcomment(mcomment);
 			
-			LogDAO ldao = new LogDAO();
-			
-			if(mode.equals("insert")) {
-				ldao.insertGrade(ldto);
-			}
-			else if(mode.equals("edit")) {
-				ldao.updateGrade(ldto);
-			}
-			else if(mode.equals("delete")) {
+			if(mode.equals("delete")) {
 				ldao.deleteGrade(ldto);
 			}
+			
+			else {
 				
+				int grade = Integer.parseInt(req.getParameter("reviewStar"));
+				String mcomment = req.getParameter("mcomment");
+				
+				ldto.setGrade(grade);
+				ldto.setMcomment(mcomment);
+				
+				if(mode.equals("insert")) {
+					ldao.insertGrade(ldto);
+				}
+				else if(mode.equals("edit")) {
+					ldao.updateGrade(ldto);
+				}
+			}
+			
 			ldao.close();
 			
 			resp.sendRedirect("./MovieViewController.do?idx="+idx);
